requirements.txt
python-dotenv>=1.0.1
spotipy>=2.23.0
requests>=2.31.0
flask>=3.0.3